Create Pivot_Count.c
Quicksort is a well known sorting algorithm. It makes use of what is called a 'pivot', which is basically an element used to restructure a given array of elements into two parts, elements smaller than the pivot and elements larger than the pivot. This algorithm is then applied recursively to these two parts till the whole array becomes sorted.

You are given an array A of size N that has already gone through the partitioning described above. Find the number of elements that could have been a pivot for this array.

Formally, an element Ai could have been a pivot if,

for every k < i, Ak <= Ai AND
for every k > i, Ak >= Ai
Constraints

1 <= N <= 105
1 <= Ai <= 109
Input Format

The first line contains a single integer N. The size of the array.

The second line contains N space-seperated integers, the array A.

Output Format

Output a single integer, the number of elements that could have been a pivot.

Sample Input

8
1 4 3 5 7 9 8 10

Sample Output

4

Explanation

The elements 1, 5, 7 and 10 could have acted as a pivot for the creation of this partitioned array.