Create sum_of_pivots.c
You are given an array A of N elements. Your job is to find the Kth smallest element in the array. If the array is A = {7,4,1,3,2} then 1 is the 0th smallest, 2 is the 1st smallest and so on.

You have to use the partitioning algorithm used in quick-sort to do this efficiently. Basically if the pivot is placed at Kth index then we have the answer,else we have to look into either the left or right depending on the index where the pivot is placed.

While solving this question you will always choose the last element of the array/sub-array as the pivot.

Formally, an element A[x] would be a pivot if,
for every i < x, A[i] < A[x] AND
for every i > x, A[i] >= A[x]
Your output will be the sum of the pivots you used while partitioning the array. Basically you have to tweak the partitioning algorithm to add the pivot it chose to compute the sum.

Input Format

First line of the input will contain N and K separated by space. Next line will contain N space separated integers.

1 <= N,K <= 105

1 <= A[i] <= 105

Note : As sum may be large, use long long or int64 specifier in C++ & long in Java to store the sum.

Output Format

Ouput a single line containing the sum of the pivots you chose.

Sample Input

5 2
4 6 3 1 7

Sample Output

12

Explanation

Step 1: 4 6 3 1 7 
Choose 7 as the pivot. Resulting array is 4 6 3 1 7 as 7 is largest element.
Step 2: 4 6 3 1 
Choose 1 as the pivot. Resulting array is 1 6 3 4 7 as 1 is the least element.
Step 3: 6 3 4 
Choose 4 as the pivot. Resulting array is 1 3 4 6 7 . Now we see A[2] = 4 and we terminate.

Sum of the pivots = 7 + 1 + 4 = 12