Create gau_aish_2.cpp
Garurav and Aishwarya have finally reached a truce. Unfortunately, the tree they lived on got destroyed in their fight. They are now looking for a new home.

All homes are in the shape of a graph. Aishwarya wants to live in a tree again. So it falls to Gaurav to determine whether the home they are currently considering is a tree or not.

Given a graph on N nodes and M edges, determine if it is a tree or not.

The nodes are numbered as follows: 0, 1, ..., N - 1.

The graph is undirected. Do not assume anything else about the graph.

1 <= N <= 100000
0 <= M <= 100000
0 <= u, v <= N - 1 for any edge (u, v)

Input Format

The first line contains two space seperated integers, N and M.
M lines follow, each containing two space seperated integers u and v signifying an edge between u and v.

Output Format

Output "YES" (without quotes) if it is a tree and "NO" otherwise.

Sample Input

3 2
1 2
0 1

Sample Output

YES

Explanation

Ob.