Create another_dijkstra
You are given an undirected weighted graph G(V, E) with N vertices and M edges.

Here V refers to the set of vertices and E refers to the set of edges.

The vertices are numbered as follow: 0, 1, ..., N - 1.

You are also given two sets A and B, both of which are a subset of V.

You are taking part in a relay race. The race requires a team of three people.

The race begins at the vertex 0, where you are standing. You can go to any one of the vertices from set A where your teammate is waiting. He then goes to any one of the vertices from set B where the third teammate is waiting. The third teammate then has to go to the finish point, vertex N - 1.

Being a dijkstra problem, you have to, of course, find the shortest path such that it satisfies the conditions of this race.

Once again, the path starts from the vertex 0. Then goes to at least one vertex from set A. Followed by at least one vertex from set B. And then finally goes to node N - 1.

You need to print the length of the shortest path.

The graph is guarranteed to be connected. The intersection of A and B is the empty set, meaning no vertex is in both A and B. Also, neither 0 nor N - 1 are a part of either A or B.

Input Format

First line contains four space seperated integers, N, M, A, B, the number of nodes, the number of edges, the number of elements in set A and the number of elements in set B respectively.
The next line contains A space seperated integers, the elements of set A.
The next line contains B space seperated integers, the elements of set B.
Next M lines contain three space seperated integers, u, v, w, indicating that there is an edge between vertices u and v with weight w.

Constraints

5 <= N <= 100000
5 <= M <= 200000
1 <= |A| <= 100000
1 <= |B| <= 100000
1 <= elements in set A and B <= N - 2
1 <= edge weights <= 1000

Output Format

Output a single integer, the required shortest path.

Sample Input

6 8 1 2
3
1 2
0 1 2
0 2 3
0 3 6
1 3 3
2 3 4
2 4 3
3 5 5
4 5 2

Sample Output

14

Explanation

0 -> 1 -> 3 -> 2 -> 4 -> 5